requirements.txt
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
gspread>=5.7.0
pandas>=2.0.0
flask>=3.0.0
beautifulsoup4>=4.12.0
python-dateutil>=2.9.0
